Move approve module to new swap package
